requirements.txt
bcrypt==3.1.4
behave==1.2.6
behave-django==1.1.0
coreapi==2.3.3
coverage==4.5.1
Django==1.11.16
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-filter==2.0
django-localflavor==1.6
django-waffle==0.14.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
easy-thumbnails==2.5
factory-boy==2.11.1
Faker==0.9.2
flake8==3.5.0
gunicorn==19.7.1
ipython==7.1.1
Markdown==2.6.11
Pillow==5.1.0
pre-commit==1.12.0
psycopg2-binary==2.7.4
pyflakes==1.6.0
python-magic==0.4.15
pytz==2018.7
redgreenunittest==0.1.1
requests==2.20.0
selenium==3.8.0
six==1.11.0
tblib==1.3.2
virtualenv==16.1.0
Werkzeug==0.14.1
